Create Java Menu
This program will greet the user and present a menu listing items such as drinks,
appetizers, and a variety of food and dessert options. With the options for the user to pick items
there will be prices attached to the food items so when an item is picked there is a monetary
value assigned to the food item. The user will pick all the food items they want until they are
content with their choices. There will be a choice for the user to exit the menu.
Once the user exits the menu, they will be shown the bill. The user will have the option to
split the bill amount multiple people if the tabs are separate. The user will enter an amount in
which they are paying and will be able to receive change for their bill. After the user pays for
their bill, they will be able to give feedback and rate their experience at the restaurant. They can
rank their experience based on food, service, and overall experience.